refactor/changed the file and package structure so that related files are stored under a single package: temp, job
