+++ resourcing-api-backend/resourcingapi/src/main/java/com/nology/HealthCheckController.java
@@ -1,0 +1,16 @@
+package com.nology;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping(value = "/")
+public class HealthCheckController {
+	
+	@GetMapping
+	public String test() {
+    	return "Hello World!!";
+    }
+
+}
