+++ resourcing-api-backend/resourcingapi/src/main/java/com/nology/ResourcingapiApplication.java
@@ -1,0 +1,16 @@
+package com.nology;
+
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ResourcingapiApplication {
+
+//	@SuppressWarnings("unchecked")
+	public static void main(String[] args) {
+		
+		SpringApplication.run(ResourcingapiApplication.class, args);
+	}
+
+}
