--- resourcing-api-backend/resourcingapi/src/main/java/com/nology/resourcingapi/ResourcingapiApplication.java
@@ -1,16 +1,0 @@
-package com.nology.resourcingapi;
-
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ResourcingapiApplication {
-
-//	@SuppressWarnings("unchecked")
-	public static void main(String[] args) {
-		
-		SpringApplication.run(ResourcingapiApplication.class, args);
-	}
-
-}
