--- resourcing-api-backend/resourcingapi/src/main/java/com/nology/resourcingapi/controller/HealthCheckController.java
@@ -1,16 +1,0 @@
-package com.nology.resourcingapi.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping(value = "/")
-public class HealthCheckController {
-	
-	@GetMapping
-	public String test() {
-    	return "Hello World!!";
-    }
-
-}
