--- resourcing-api-backend/resourcingapi/src/main/java/com/nology/resourcingapi/dto/TempCreateDTO.java
@@ -1,22 +1,0 @@
-package com.nology.resourcingapi.dto;
-
-import javax.validation.constraints.NotNull;
-
-
-public class TempCreateDTO {
-	
-	@NotNull
-	private String firstName;
-	
-	@NotNull
-	private String lastName;
-	
-	public String getFirstName() {
-		return firstName;
-	}
-	
-	public String getLastName() {
-		return lastName;
-	}
-	
-}
