--- resourcing-api-backend/resourcingapi/src/main/java/com/nology/resourcingapi/exception/FieldInvalidError.java
@@ -1,19 +1,0 @@
-package com.nology.resourcingapi.exception;
-
-public class FieldInvalidError {
-	
-    private String message;
-
-    public FieldInvalidError(String message) {
-        this.message = message;
-    }
-
-	public String getMessage() {
-		return message;
-	}
-
-	public void setMessage(String message) {
-		this.message = message;
-	}
-    
-}
