--- resourcing-api-backend/resourcingapi/src/main/java/com/nology/resourcingapi/exception/ResourceNotFoundException.java
@@ -1,10 +1,0 @@
-package com.nology.resourcingapi.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-
-	  private static final long serialVersionUID = 1L;
-
-	  public ResourceNotFoundException(String msg) {
-	    super(msg);
-	  }
-	}
