+++ resourcing-api-backend/resourcingapi/src/main/java/com/nology/temp/TempCreateDTO.java
@@ -1,0 +1,22 @@
+package com.nology.temp;
+
+import javax.validation.constraints.NotNull;
+
+
+public class TempCreateDTO {
+	
+	@NotNull
+	private String firstName;
+	
+	@NotNull
+	private String lastName;
+	
+	public String getFirstName() {
+		return firstName;
+	}
+	
+	public String getLastName() {
+		return lastName;
+	}
+	
+}
