+++ resourcing-api-backend/resourcingapi/src/main/java/exceptions/FieldInvalidError.java
@@ -1,0 +1,19 @@
+package exceptions;
+
+public class FieldInvalidError {
+	
+    private String message;
+
+    public FieldInvalidError(String message) {
+        this.message = message;
+    }
+
+	public String getMessage() {
+		return message;
+	}
+
+	public void setMessage(String message) {
+		this.message = message;
+	}
+    
+}
