+++ resourcing-api-backend/resourcingapi/src/main/java/exceptions/ResourceNotFoundException.java
@@ -1,0 +1,10 @@
+package exceptions;
+
+public class ResourceNotFoundException extends RuntimeException {
+
+	  private static final long serialVersionUID = 1L;
+
+	  public ResourceNotFoundException(String msg) {
+	    super(msg);
+	  }
+	}
